Delete previous create_new_parent
